% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbiobank.R
\name{check_for_code}
\alias{check_for_code}
\title{Check for one or several codes in all columns of a dataset}
\usage{
check_for_code(codes, col_name, data)
}
\arguments{
\item{codes}{a vector of codes, either character or numeric.}

\item{col_name}{Should all individuals meeting the criteria be marked with 1 in a column?}

\item{data}{the UKB data, read into memory with \code{\link[=ukb_extract]{ukb_extract()}}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Check for one or several codes in all columns of a dataset
}
\examples{
\dontrun{
bipolar <- check_for_code(c("F30", "F31", "bipolar", icd_data))
}
}
