% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectortools.R
\name{colRep}
\alias{colRep}
\title{rep a colVector to a matrix}
\usage{
colRep(colWert, dim2)
}
\arguments{
\item{colWert}{numic vector}

\item{dim2}{intger, number of cols}
}
\value{
a matrix
}
\description{
rep a colVector to a matrix
}
