% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime3.R
\name{crime3}
\alias{crime3}
\title{Access one of the tables in crime3}
\usage{
crime3(table, prefix = "V_\%")
}
\arguments{
\item{table}{Name of the table. Call crime3() without arguments to see a
list of all tables}

\item{prefix}{By default, calling crime3 without arguments lists all tables
starting with 'V_'*. By specifiying another prefix, other tables can be listed.}
}
\value{
a dblyr table, or character vector. To further get the data into the R
session, collect() needs to be called.
}
\description{
Access one of the tables in crime3
}
\examples{
\dontrun{
available_tables <- crime3()
meds2005 <- crime3("V_LMED2005")
}
}
