% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime3.R
\name{crime3_con}
\alias{crime3_con}
\title{Establish connection to crime3}
\usage{
crime3_con()
}
\value{
An odbc connection object
}
\description{
Establish connection to crime3
}
\examples{
con <- crime3_con()
}
