% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime3.R
\name{crime3_first_unique_diagnosis}
\alias{crime3_first_unique_diagnosis}
\title{Finds the first date for each unique diagnosis}
\usage{
crime3_first_unique_diagnosis(tbl, icd_digits = 3)
}
\arguments{
\item{tbl}{a tibble from crime3_icd()}

\item{icd_digits}{How many digits of the icd code to consider as a unique diagnosis.
3 digits would cut F323A at F32, while four digits would allow for distinction between
F323 and F322}
}
\value{
a tibble
}
\description{
Finds the first date for each unique diagnosis
}
\examples{
\dontrun{
mdd <- crime3_icd(c("F32", "F33"))
first_dia <- crime3_first_unique_diagnosis(mdd, 3)
}
}
