% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime3.R
\name{crime3_icd}
\alias{crime3_icd}
\title{Filter NPR or GVR for individuals with a set of icd codes}
\usage{
crime3_icd(code, specialist = TRUE)
}
\arguments{
\item{code}{a character vector of ICD codes}

\item{specialist}{a boolean, TRUE to look in NPR, false to look in primary care}
}
\value{
a tibble of individuals
}
\description{
Filter NPR or GVR for individuals with a set of icd codes
}
\examples{
\dontrun{
mdd <- crime_icd(c("F32", "F33"))
mdd_primary_care <- crime_icd(c("F32", "F33"), specialist=FALSE)
}
}
