% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectortools.R
\name{fctPiecewiseFunction}
\alias{fctPiecewiseFunction}
\title{function for a 2 Piecewise}
\usage{
fctPiecewiseFunction(fct1, fct2, Vectorr, judge)
}
\arguments{
\item{fct1}{Function, first function}

\item{fct2}{Function, second function}

\item{Vectorr}{Vector, input(x)}

\item{judge}{numic, judge wert}
}
\value{
Result output(y)
}
\description{
function for a 2 Piecewise
}
