% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_date.R
\name{first_date}
\alias{first_date}
\title{Title}
\usage{
first_date(tbl, datecols)
}
\arguments{
\item{tbl}{a tibble,}

\item{datecols}{name of columns that you want to find the earliest date among}
}
\value{
a tibble with a new column: first_date, which is the first date of all
date columns or among datecols
}
\description{
Title
}
\examples{
\dontrun{
mdd <- crime3_icd(c("F32", "F33"))
first_dia <- crime3_first_unique_diagnosis(mdd, 3)
first_date(first_dia)
}
}
