% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime3.R
\name{get_icd_code}
\alias{get_icd_code}
\title{Access the national patient registry, or outpatient registry care for stockholm}
\usage{
get_icd_code(code, specialist = TRUE)
}
\arguments{
\item{code}{Filter individuals with diagnosis starting with the code. Specifying F32 will
filter all individuals with F32*. Assumes you have defined con with
con <- crime3_con()}

\item{specialist}{Use specialist care, or primary care for stockholm county?}
}
\value{
a dbplyr table
}
\description{
Access the national patient registry, or outpatient registry care for stockholm
}
\examples{
\dontrun{
mdd <- get_icd_code("F32")
scz <- get_icd_code("F25")
}
}
