% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectortools.R
\name{integrateMAP2}
\alias{integrateMAP2}
\title{Adaptive quadrature of functions of one variable over a finite or infinite interval.}
\usage{
integrateMAP2(lower, upper, f, ...)
}
\arguments{
\item{lower}{the limits of integration. Can be infinite.}

\item{upper}{the limits of integration. Can be infinite.}

\item{f}{an R function taking a numeric first argument and returning a numeric vector of the same length. 
Returning a non-finite element will generate an error.}

\item{...}{additional arguments to be passed to f.}
}
\value{
value   the final estimate of the integral.
}
\description{
Adaptive quadrature of functions of one variable over a finite or infinite interval.
}
