% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectortools.R
\name{integrateVector}
\alias{integrateVector}
\title{Adaptive quadrature of functions of one variable over a finite or infinite interval.}
\usage{
integrateVector(f, lower, upper, ...)
}
\arguments{
\item{f}{an R function taking a numeric first argument and returning a numeric vector of the same length. 
Returning a non-finite element will generate an error.}

\item{lower}{vector the limits of integration. Can be infinite.}

\item{upper}{vector the limits of integration. Can be infinite.}

\item{...}{additional arguments to be passed to f.}
}
\value{
value a vector the final estimate of the integral.
}
\description{
Adaptive quadrature of functions of one variable over a finite or infinite interval.
}
