% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectortools.R
\name{makeLTM}
\alias{makeLTM}
\title{judge location translate matrix}
\usage{
makeLTM(Vectorr, lengthV, judge)
}
\arguments{
\item{Vectorr}{Vector, input(x)}

\item{lengthV}{forget}

\item{judge}{numic, judge wert}
}
\value{
2-matrix of location
}
\description{
judge location translate matrix
}
