% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectortools.R
\name{maxSVector}
\alias{maxSVector}
\title{Finding the maximum value of one vectors and one wert}
\usage{
maxSVector(WrtA, VctB)
}
\arguments{
\item{WrtA}{numic}

\item{VctB}{vector, numic}
}
\value{
vector, result
}
\description{
Finding the maximum value of one vectors and one wert
}
\examples{
a = 2
B = c(1, 2, 3)
maxVector(a, B)
}
