% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectortools.R
\name{maxVector}
\alias{maxVector}
\title{Finding the maximum value of the corresponding position of two equal large vectors}
\usage{
maxVector(VctA, VctB)
}
\arguments{
\item{VctA}{vector, numic, first Vector}

\item{VctB}{vector, numic, second Vector}
}
\value{
vector, result
}
\description{
Finding the maximum value of the corresponding position of two equal large vectors
}
\examples{
A = c(1, 2, 3)
B = c(2, 2, 2)
maxVector(A, B)
}
