% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych4.R
\name{psych4}
\alias{psych4}
\title{Finds the table in the psych4 data}
\usage{
psych4(table, prefix = "V_\%")
}
\arguments{
\item{table}{\itemize{
\item the table you want to access
}}

\item{prefix}{\itemize{
\item for filtering to other tables, used by DB admins
}}
}
\value{
a dbplyr connection object
}
\description{
Finds the table in the psych4 data
}
\examples{
\dontrun{
psych4("V_NPR_DIA")
# to list all tables starting with V_ (those that have actual data for us)
psych4()
}
}
