% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych4.R
\name{psych4_con}
\alias{psych4_con}
\title{Establish connection to crime3}
\usage{
psych4_con()
}
\value{
An odbc connection object
}
\description{
Establish connection to crime3
}
\examples{
\dontrun{
con <- psych4()
}
}
