% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbiobank.R
\name{ukb_extract}
\alias{ukb_extract}
\title{Extract a set of fields for UKBB data on Vector}
\usage{
ukb_extract(variables, ukb_data, remove_withdrawn = TRUE)
}
\arguments{
\item{variables}{Character vector of field names}

\item{ukb_data}{Path to UKBB data. If not passed will scan all dataframes in the UKBB folder}

\item{remove_withdrawn}{Remove participants that has withdrawn their consent?}
}
\value{
a tibble of the the datafields extracted
}
\description{
Extract a set of fields for UKBB data on Vector
}
\examples{
\dontrun{
df <- ukb_extract(c("20002", "41270"))
}
}
