% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbiobank.R
\name{check_for_code}
\alias{check_for_code}
\title{Check for one or several codes in all columns of a dataset}
\usage{
check_for_code(codes, col_name, data)
}
\arguments{
\item{codes}{codes to check for}

\item{col_name}{name of the column to be created}

\item{data}{a dataframe}
}
\value{
a tibble
}
\description{
Check for one or several codes in all columns of a dataset
}
\examples{
\dontrun{
bipolar <- check_for_code(c("F30", "F31", "bipolar", icd_data))
}
}
