% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbiobank.R
\name{ukb_extract}
\alias{ukb_extract}
\title{Title}
\usage{
ukb_extract(variables, ukb_data, remove_withdrawn = TRUE)
}
\arguments{
\item{variables}{vector or string of column names}

\item{ukb_data}{list of filepaths for the UKB data files}

\item{remove_withdrawn}{Remove participants that has withdrawn their consent?}
}
\value{
a tibble of the the datafields extracted
}
\description{
Title
}
\examples{
\dontrun{
df <- ukb_extract(c("20002", "41270"))
}
}
